make crate into a library
This allows interested parties to consume code in
here for their own application, as well as the
addition of integration tests in an upcoming change.
